+++ src/db/sequelize-base.service.ts
@@ -14,7 +14,7 @@ import {
 } from '../gql';
 import { calculateBeforeAndAfter, calculateLimitAndOffset } from './index';
 
-import { Model } from 'sequelize-typescript';
+import { Model, DataType } from 'sequelize-typescript';
 
 import { toBase64 } from '../helpers/base64';
 import { ClassType } from '../helpers/classtype';
@@ -26,7 +26,7 @@ import {
 } from './db-to-gql';
 import { Context } from '../server/index';
 import { publishCurrentState } from './gql-pubsub-sequelize-engine';
-import { Transaction, FindOptions, Op, Order } from 'sequelize';
+import { Transaction, FindOptions, Op, Order, OrderItem } from 'sequelize';
 import { findEach } from 'iterable-model';
 import { Actions, RFIAuthError, Permissions, AuthorizerTreatAsMap, Scopes } from '@rumbleship/acl';
 import { createWhereClauseWith } from '../gql/create-where-clause-with';
@@ -434,6 +434,20 @@ export class SequelizeBaseService<
     //
     const connection = new this.connectionClass();
 
+    const id_attribute = this.model.rawAttributes['id'];
+    let orderClause: OrderItem[] | undefined;
+    if (id_attribute) {
+      // if it is a uuid or string as id, then we want to use created_at if its there...
+      // but we may not have either condition, in which case there is no default ordering
+      if (id_attribute.type === DataType.NUMBER) {
+        orderClause = [['id', 'DESC']];
+      } else {
+        const created_at_attribute = this.model.rawAttributes['id'];
+        if (created_at_attribute) {
+          orderClause = [['created_at', 'DESC']];
+        }
+      }
+    }
     const limits = calculateLimitAndOffset(after, first, before, last);
     const whereClause = createWhereClauseWith(filter);
     const sequelizeOptions = this.convertServiceOptionsToSequelizeOptions(options) ?? {};
@@ -441,6 +455,7 @@ export class SequelizeBaseService<
       where: whereClause,
       offset: limits.offset,
       limit: limits.limit,
+      order: orderClause,
       ...sequelizeOptions
     };